requirements.txt
streamlit==1.32.0
pandas==2.2.1
requests==2.31.0
python-dotenv==1.0.1
scikit-learn==1.4.1.post1
urllib3==2.2.1
setuptools==69.1.1
pip>=24.0